requirements.txt
tabulate
python-magic-bin>=0.4.14
